long long -> int64_t, for CRAN's sake

+++ src/gengraph_graph_molloy_hash.cpp
@@ -906,9 +906,9 @@ bool graph_molloy_hash::make_connected() {
   return(trees == ffub);
 }
 
-long long graph_molloy_hash::slow_connected_shuffle(long long times) {
+int64_t graph_molloy_hash::slow_connected_shuffle(int64_t times) {
   assert(verify());
-  long long nb_swaps = 0;
+  int64_t nb_swaps = 0;
   int T = 1;
 
   while(times>nb_swaps) {
